Update README.md
# PROG_PART2
PART2
MY APP IS VERY EASY TOO USE IT HAS DIFFERENT ESPECTS ON WHAT TO EXPECT WHEN ENTERING THE INFORMATION
This web application is a a recipe app that allows you to record your ingredients, type of food the ingredient is, the calories that it requires , the scale of the calories that are needed for the ingredients that you have recorded in the web application. You can also reset if you have made an error while recording the data of the ingredients. The app makes calories of your ingredients not be able exceed 300 but it also allows you too enter unlimited ingredients